Move useTimeout from common utils to controllers/hooks
