+++ src/controllers/index.ts
@@ -11,6 +11,7 @@ export {
   useRouterActions,
   useQueryParam,
   usePathParam,
+  useTimeout,
 } from './hooks';
 export {
   useResourceStoreContext,
